Correctly handle all-day events
An inclusive range for multi-day all-day events seems to work much
better than a range exclusive at the upper end, which is what Google
Calendar returns by default.  With the exclusive range, I tried a few
different prompts but the LLM wasn't quite sure whether the end date was
to be included or not.